day-07 nice, bump many things to int64

+++ 2024/Day.fs
@@ -3,9 +3,9 @@
 type Day =
     {
         dayNumber : int
-        solvePart0 : string list -> int
-        solvePart1 : string list -> int
-        inputs : (string * (int * int) option) list
+        solvePart0 : string list -> int64
+        solvePart1 : string list -> int64
+        inputs : (string * (int64 * int64) option) list
     }
 
 let day number solvePart0 solvePart1 =
